update spacing builder
