+++ resources/grammar/ObjJ.bnf
@@ -215,8 +215,7 @@ private sourceElement
 	| 	protocolDeclaration
 	| 	preprocessorStatement
 	| 	functionDeclaration
-	|	importFile
-    | 	importFramework
+	|	importBlock
 	|	globalVariableDeclaration eos?
 	|	expr eos?
     | 	bodyVariableAssignment eos?
@@ -229,8 +228,7 @@ private sourceElement
 	| 	global
 	| 	classDependencyStatement
 	|   comment
-	|	includeFile
-	|	includeFramework
+	|	includeBlock
 	| 	emptyStatement
 	|	fragment eos?
 	;
@@ -260,6 +258,15 @@ blockCommentBody
 	| BLOCK_COMMENT_LINE
 	;
 
+importBlock
+	::= importStatementElement+
+	;
+
+importStatementElement
+	::= importFile
+	|	importFramework
+	;
+
 importFile
 	::= '@import' stringLiteral
 	{
@@ -278,6 +285,16 @@ importFramework
 		stubClass="cappuccino.ide.intellij.plugin.stubs.interfaces.ObjJImportStub<cappuccino.ide.intellij.plugin.psi.impl.ObjJImportFrameworkImpl>"
 		methods=[getFrameworkName getFileName getImportAsUnifiedString]
 	}
+	;
+
+
+includeBlock
+	::= includeStatementElement+
+	;
+
+includeStatementElement
+	::= includeFile
+	|	includeFramework
 	;
 
 includeFile
@@ -340,7 +357,7 @@ private preprocessorBodyStatement
 	::= <<inMode "inImplementation" >> methodDeclaration
     |	<<inMode "inProtocol" >> protocolScopedMethodBlock
     |	<<inMode "inProtocol" >> protocolHeader
-    |  	<<inMode "implementationInstanceVariables">> instanceVariableDeclaration eos?
+    |  	<<inMode "implementationInstanceVariables">> instanceVariableDeclaration
     |	<<notInMode "inProtocol">> <<notInMode "inImplementation">> sourceElement
     |	elseIfStatement
     |	elseStatement
@@ -481,7 +498,7 @@ variableDeclaration
 	;
 
 instanceVariableDeclaration
-	::= ('@outlet')? formalVariableType variableName accessor?
+	::= ('@outlet')? formalVariableType variableName accessor? eos
 	{
 		extends="cappuccino.ide.intellij.plugin.psi.impl.ObjJStubBasedElementImpl<?>"
 		stubClass="cappuccino.ide.intellij.plugin.stubs.interfaces.ObjJInstanceVariableDeclarationStub"
@@ -489,7 +506,6 @@ instanceVariableDeclaration
 			"cappuccino.ide.intellij.plugin.psi.interfaces.ObjJHasTreeStructureElement"
 			"cappuccino.ide.intellij.plugin.psi.interfaces.ObjJHasContainingClass"
 			"cappuccino.ide.intellij.plugin.psi.interfaces.ObjJResolveableElement<cappuccino.ide.intellij.plugin.stubs.interfaces.ObjJInstanceVariableDeclarationStub>"
-			"cappuccino.ide.intellij.plugin.psi.interfaces.ObjJNeedsSemiColon"
 		]
 		methods = [getGetter getSetter getName setName hasContainingClass shouldResolve getContainingSuperClassName getContainingSuperClass getContainingClass getContainingClassName createTreeStructureElement getAccessorPropertyList]
 		pin=2
@@ -1011,7 +1027,7 @@ instanceVariableList
 	;
 
 private instanceVariableListList
-	::= ( comment | preprocessorInstanceVariableListIfStatement | instanceVariableDeclaration eos? )+
+	::= ( comment | preprocessorInstanceVariableListIfStatement | instanceVariableDeclaration)+
 	{
 		recoverWhile=recover_instanceVariableList
 	}
@@ -1125,7 +1141,7 @@ private recover_preprocessorIfCondition
 	;
 
 private preprocessorInstanceVariableList
-	::= (comment|instanceVariableDeclaration eos?|LINE_TERMINATOR)+
+	::= (comment|instanceVariableDeclaration|LINE_TERMINATOR)+
 	{
 		recoverWhile=recover_preprocessorIfCondition
 	}
