+++ src/cappuccino/ide/intellij/plugin/psi/types/ObjJTokenSets.kt
@@ -10,7 +10,7 @@ import com.intellij.psi.tree.TokenSet.create
 object ObjJTokenSets {
 
     val BLOCKS = create(ObjJ_BLOCK_ELEMENT, ObjJ_BRACKET_LESS_BLOCK, ObjJ_METHOD_BLOCK, ObjJ_PROTOCOL_SCOPED_METHOD_BLOCK,
-            ObjJ_STATEMENT_OR_BLOCK)
+            ObjJ_STATEMENT_OR_BLOCK , ObjJ_INSTANCE_VARIABLE_LIST)
     val BLOCKS_EXT = create(ObjJ_BLOCK_ELEMENT, ObjJ_BRACKET_LESS_BLOCK, ObjJ_METHOD_BLOCK, ObjJ_PROTOCOL_SCOPED_METHOD_BLOCK,
             ObjJ_STATEMENT_OR_BLOCK, ObjJStubTypes.FILE)
 
@@ -22,6 +22,8 @@ object ObjJTokenSets {
     val STRING_LITERALS = create(ObjJ_SINGLE_QUOTE_STRING_LITERAL, ObjJ_DOUBLE_QUOTE_STRING_LITERAL)
 
     val IMPORT_STATEMENTS = create(ObjJ_IMPORT_FILE, ObjJ_IMPORT_FRAMEWORK)
+
+    val IMPORT_BLOCKS = create(ObjJ_IMPORT_BLOCK, ObjJ_INCLUDE_BLOCK);
 
     val FUNCTION_DECLARATIONS = create(ObjJ_FUNCTION_DECLARATION, ObjJ_FUNCTION_LITERAL, ObjJ_METHOD_BLOCK)
 
@@ -48,6 +50,28 @@ object ObjJTokenSets {
             ObjJ_BODY_VARIABLE_ASSIGNMENT,
             ObjJ_BLOCK_ELEMENT,
             ObjJ_COMMENT
+    )
+
+    val EXPRESSIONS = create(
+            ObjJ_RETURN_STATEMENT,
+            ObjJ_THROW_STATEMENT,
+            ObjJ_DEBUGGER_STATEMENT,
+            ObjJ_DELETE_STATEMENT,
+            ObjJ_EXPR,
+            ObjJ_BODY_VARIABLE_ASSIGNMENT
+    )
+
+    val NEW_LINE_AFTER = create(
+            ObjJ_IF_STATEMENT,
+            ObjJ_FOR_STATEMENT,
+            ObjJ_SWITCH_STATEMENT,
+            ObjJ_TRY_STATEMENT,
+            ObjJ_FUNCTION_DECLARATION,
+            ObjJ_IMPORT_BLOCK,
+            ObjJ_INCLUDE_BLOCK,
+            ObjJ_BLOCK_ELEMENT,
+            ObjJ_DO_WHILE_STATEMENT,
+            ObjJ_FUNCTION_LITERAL
     )
 
     val LOGIC_OPERATORS = create(
@@ -155,25 +179,32 @@ object ObjJTokenSets {
     )
 
     val VAR_TYPE_KEYWORDS = create(
-            ObjJTypes.ObjJ_VAR_TYPE_BOOL,
-            ObjJTypes.ObjJ_VAR_TYPE_INT,
-            ObjJTypes.ObjJ_VAR_TYPE_SHORT,
-            ObjJTypes.ObjJ_VAR_TYPE_LONG,
-            ObjJTypes.ObjJ_VAR_TYPE_LONG_LONG,
-            ObjJTypes.ObjJ_VAR_TYPE_UNSIGNED,
-            ObjJTypes.ObjJ_VAR_TYPE_SIGNED,
-            ObjJTypes.ObjJ_VAR_TYPE_FLOAT,
-            ObjJTypes.ObjJ_VAR_TYPE_DOUBLE,
-            ObjJTypes.ObjJ_VAR_TYPE_BYTE,
-            ObjJTypes.ObjJ_VAR_TYPE_ID
+            ObjJ_VAR_TYPE_BOOL,
+            ObjJ_VAR_TYPE_INT,
+            ObjJ_VAR_TYPE_SHORT,
+            ObjJ_VAR_TYPE_LONG,
+            ObjJ_VAR_TYPE_LONG_LONG,
+            ObjJ_VAR_TYPE_UNSIGNED,
+            ObjJ_VAR_TYPE_SIGNED,
+            ObjJ_VAR_TYPE_FLOAT,
+            ObjJ_VAR_TYPE_DOUBLE,
+            ObjJ_VAR_TYPE_BYTE,
+            ObjJ_VAR_TYPE_ID
     )
 
     val ITERATION_STATEMENT_KEYWORDS = create(
-            ObjJTypes.ObjJ_IF,
-            ObjJTypes.ObjJ_ELSE,
-            ObjJTypes.ObjJ_IN,
-            ObjJTypes.ObjJ_FOR,
-            ObjJTypes.ObjJ_WHILE,
-            ObjJTypes.ObjJ_DO
+            ObjJ_IF,
+            ObjJ_ELSE,
+            ObjJ_IN,
+            ObjJ_FOR,
+            ObjJ_WHILE,
+            ObjJ_DO
+    )
+
+    val INDENT_FIRST_CHILD = create(
+            ObjJ_DO_WHILE_STATEMENT,
+            ObjJ_WHILE_STATEMENT,
+            ObjJ_IF_STATEMENT,
+            ObjJ_FOR_STATEMENT
     )
 }
