+++ src/cappuccino/ide/intellij/plugin/psi/utils/ObjJTreeUtil.kt
@@ -6,7 +6,6 @@ import cappuccino.ide.intellij.plugin.psi.types.ObjJTokenSets
 import com.intellij.lang.ASTNode
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiErrorElement
-import com.intellij.psi.impl.source.tree.TreeUtil
 import com.intellij.psi.stubs.StubElement
 import com.intellij.psi.tree.IElementType
 import com.intellij.psi.util.PsiTreeUtil
@@ -17,6 +16,11 @@ import com.intellij.openapi.editor.Editor
 import com.intellij.psi.TokenType
 
 import java.util.ArrayList
+import java.util.logging.Logger
+
+private val LOGGER:Logger by lazy {
+    Logger.getLogger("#ObjJTreeUtil")
+}
 
 
 fun PsiElement?.getChildrenOfType(iElementType: IElementType): List<PsiElement> {
@@ -104,10 +108,10 @@ fun ASTNode.getPreviousNonEmptyNodeIgnoringComments(): ASTNode? {
     return node
 }
 fun ASTNode?.getPreviousNonEmptyNode(ignoreLineTerminator: Boolean): ASTNode? {
-    var out: ASTNode? = this?.treePrev ?: this?.treeParent?.treePrev ?: return null
+    var out: ASTNode? = this?.treePrev ?: getPrevInTreeParent(this) ?: return null
     while (out != null && shouldSkipNode(out, ignoreLineTerminator)) {
         out = if (out.treePrev == null) {
-            out.treeParent?.treePrev
+            getPrevInTreeParent(out)
         } else {
             out.treePrev
         }
@@ -117,6 +121,14 @@ fun ASTNode?.getPreviousNonEmptyNode(ignoreLineTerminator: Boolean): ASTNode? {
         //LOGGER.log(Level.INFO, "<"+compositeElement.getText()+">NextNode "+foldingDescriptors.getText()+" ElementType is <"+foldingDescriptors.getElementType().toString()+">");
     }
     return out
+}
+
+private fun getPrevInTreeParent(out:ASTNode?): ASTNode? {
+    var temp:ASTNode? = out?.treeParent ?: return null
+    while (temp != null && temp.treePrev == null && temp.treeParent != null) {
+        temp = temp.treeParent
+    }
+    return temp?.treePrev
 }
 
 
@@ -135,19 +147,26 @@ fun ASTNode.getNextNonEmptyNodeIgnoringComments(): ASTNode? {
     return node
 }
 fun ASTNode?.getNextNonEmptyNode(ignoreLineTerminator: Boolean): ASTNode? {
-    var out: ASTNode? = this?.treeNext ?: this?.treeParent?.treeNext ?: return null
+    var out: ASTNode? = this?.treeNext ?: getNextInTreeParent(this) ?: return null
     while (out != null && shouldSkipNode(out, ignoreLineTerminator)) {
         out = if (out.treeNext == null) {
-            out.treeParent?.treeNext
+            getNextInTreeParent(out)
         } else {
             out.treeNext
         }
         if (out == null) {
             return null
         }
-        //LOGGER.log(Level.INFO, "<"+compositeElement.getText()+">NextNode "+foldingDescriptors.getText()+" ElementType is <"+foldingDescriptors.getElementType().toString()+">");
     }
     return out
+}
+
+private fun getNextInTreeParent(out:ASTNode?): ASTNode? {
+    var temp:ASTNode? = out?.treeParent ?: return null
+    while (temp != null && temp.treeNext == null && temp.treeParent != null) {
+        temp = temp.treeParent
+    }
+    return temp?.treeNext
 }
 
 fun PsiElement.getNextNonEmptySibling(ignoreLineTerminator: Boolean): PsiElement? {
@@ -188,8 +207,11 @@ fun PsiElement.distanceFromStartOfLine(document:Document) : Int {
 fun ASTNode.isDirectlyPrecededByNewline(): Boolean {
     var node:ASTNode? = this.treePrev
     while (node != null) {
+        if (node.elementType == ObjJTypes.ObjJ_LINE_TERMINATOR)
+            return true
         if (node.elementType == TokenType.WHITE_SPACE) {
-            if (node.text.contains("\n")) return true
+            if (node.text.contains("\n"))
+                return true
             node = node.treePrev
             continue
         }
@@ -268,5 +290,14 @@ fun <StubT : StubElement<*>> filterStubChildren(children: List<StubElement<*>>?,
 
 
 internal fun shouldSkipNode(out: ASTNode?, ignoreLineTerminator: Boolean): Boolean {
-    return out != null && ((ignoreLineTerminator && out.elementType === ObjJTypes.ObjJ_LINE_TERMINATOR) || out.elementType === com.intellij.psi.TokenType.WHITE_SPACE || out.psi is PsiErrorElement)
+    if (out == null) {
+        return false
+    }
+    return if (ignoreLineTerminator && out.elementType === ObjJTypes.ObjJ_LINE_TERMINATOR)
+        true
+    else if (!ignoreLineTerminator && out.text.contains("\n")) {
+        false
+    } else {
+        out.elementType === TokenType.WHITE_SPACE || out.psi is PsiErrorElement
+    }
 }
